% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undpcomposite.R
\docType{data}
\name{undpcomposite}
\alias{undpcomposite}
\title{undpcomposite: Composite data from UNDP}
\format{

}
\usage{
undpcomposite
}
\description{
Composite Data from UNDP on all indices in the form of a timeseries.
}
\keyword{datasets}
